+++ Biblioteca/src/formularios/UsuariosCadastrados.java
@@ -6,7 +6,9 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import javax.swing.JLabel;
 import javax.swing.JOptionPane;
+import javax.swing.JPasswordField;
 import javax.swing.table.DefaultTableModel;
 import model.bean.Usuarios;
 import model.dao.UsuariosDAO;
@@ -21,6 +23,10 @@ import net.proteanit.sql.DbUtils;
  * @author lyand
  */
 public class UsuariosCadastrados extends javax.swing.JInternalFrame {
+
+    private boolean autenticacao;
+    private String senhaDigitada;
+    JPasswordField pass = new JPasswordField();
 
     public UsuariosCadastrados() {
         initComponents();
@@ -90,6 +96,28 @@ public class UsuariosCadastrados extends javax.swing.JInternalFrame {
         } finally {
             ConnectionFactory.closeConnection(con, stmt, rs);
         }
+    }
+
+    public boolean verificarSenha() {
+        Connection con = ConnectionFactory.getConnection();
+        PreparedStatement stmt = null;
+        ResultSet rs = null;
+        String sql = "SELECT senha FROM tbl_admins WHERE senha LIKE ?";
+
+        try {
+            stmt = con.prepareStatement(sql);
+            stmt.setString(1, senhaDigitada);
+            rs = stmt.executeQuery();
+
+            if (rs.next()) {
+                autenticacao = true;
+            }
+        } catch (SQLException ex) {
+            JOptionPane.showMessageDialog(null, "Erro ao validar senha de usuário: " + ex);
+        } finally {
+            ConnectionFactory.closeConnection(con, stmt, rs);
+        }
+        return autenticacao;
     }
 
     @SuppressWarnings("unchecked")
@@ -212,14 +240,31 @@ public class UsuariosCadastrados extends javax.swing.JInternalFrame {
     //Exclui usuários.
     private void excluirUsuarioActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_excluirUsuarioActionPerformed
         if (tblConsultUser.getSelectedRow() != -1) {
-            Usuarios users = new Usuarios();
-            UsuariosDAO dao = new UsuariosDAO();
+            int input = JOptionPane.showConfirmDialog(null, "Para excluir um usuário, é necessário confirmação com senha", "Excluir Usuário", JOptionPane.YES_NO_OPTION, JOptionPane.INFORMATION_MESSAGE);
 
-            users.setId((int) tblConsultUser.getValueAt(tblConsultUser.getSelectedRow(), 0));
+            if (input == 0) {
+                JLabel digSenha = new JLabel("Digite sua senha de usuário:");
+                JOptionPane.showConfirmDialog(null,
+                        new Object[]{digSenha, pass}, "Senha",
+                        JOptionPane.OK_CANCEL_OPTION);
 
-            dao.delete(users);
+                senhaDigitada = new String(pass.getPassword());
 
-            readJTable();
+                verificarSenha();
+
+                if (autenticacao == true) {
+                    Usuarios users = new Usuarios();
+                    UsuariosDAO dao = new UsuariosDAO();
+
+                    users.setId((int) tblConsultUser.getValueAt(tblConsultUser.getSelectedRow(), 0));
+
+                    dao.delete(users);
+
+                    readJTable();
+                } else {
+                    JOptionPane.showMessageDialog(null, "Autorização inválida. Verifique os dados inseridos.");
+                }
+            }
         } else {
             JOptionPane.showMessageDialog(null, "Selecione um usuário para excluir.");
         }
